Add Enemy struct and functionality to the game
Introduced the Enemy struct with properties, animations, and behavior methods (init, draw, update). Integrated enemies into the game loop and updated CMakeLists to include the new Enemy module. Fixed typecasting issue in Player animation timing.

+++ Player/Player.c
@@ -100,7 +100,7 @@ void UpdatePlayer(struct Player *player, const float deltaTime)
     player->frameTimer += deltaTime;
 
     // Convert frameSpeed to the time each frame should be displayed
-    float frameDuration = 1.0f / player->frameSpeed;
+    float frameDuration = 1.0f / (float) player->frameSpeed;
 
     // Check if it's time for the next frame
     if (player->frameTimer >= frameDuration) {
